refinement and add more doc strings
