+++ lex_test.go
@@ -431,7 +431,7 @@ func TestLex(t *testing.T) {
 			defer file.Close()
 			options := LexOptions{
 				ExternalLexers: []ExternalLexer{
-					&LuaLexer{},
+					&Lua{},
 				},
 			}
 			i := 0
